persistent cobrasync (#1612)
* update sync script to allow commits without needing to remove the vitess clone


* allow sync_cobradocs to be run multiple times with the same vitess clone


---------

+++ tools/sync_cobradocs.sh
@@ -2,20 +2,27 @@
 
 set -exuo pipefail
 
-tmp="$(mktemp -d vitessio.website)"
-trap "rm -rf ${tmp}" EXIT
+persist_tmpdir=${COBRADOCS_SYNC_PERSIST:-}
+tmp="vitessio.website"
 
-git clone --depth=1 git@github.com:vitessio/vitess "${tmp}/vitess" && \
-    cd "${tmp}/vitess" && \
-    git fetch --tags && \
-    cd -
+if [[ -z "${persist_tmpdir}" ]]; then
+    trap "rm -rf ${tmp}" EXIT
+fi
+
+if [ ! -d "${tmp}" ]; then
+    tmp="$(mktemp -d vitessio.website)"
+
+    git clone --depth=1 git@github.com:vitessio/vitess "${tmp}/vitess" && \
+        cd "${tmp}/vitess" && \
+        git fetch --tags && \
+        cd -
+fi
 
 VITESS_DIR="$(pwd)/${tmp}/vitess" make generated-docs
-rm -rf "${tmp}/vitess"
 
 git add $(git diff -I"^commit:.*$" --numstat | awk '{print $3}' | xargs)
 # Reset any modified files that contained _only_ a SHA update.
 git checkout -- .
 # Add any net-new files that were missed in the first `git add`.
-git add .
+git add content/**
 git commit -sm "sync cobradocs to latest release branches"
